+++ Editor/Package.cs
@@ -2,11 +2,12 @@
 using System;
 using UnityEditor;
 
-namespace Hananoki.BuildAssist {
+namespace HananokiEditor.BuildAssist {
   public static class Package {
     public const string name = "BuildAssist";
+    public const string nameNicify = "Build Assist";
     public const string editorPrefName = "Hananoki.BuildAssist";
-    public const string version = "1.3.6";
+    public const string version = "1.3.7";
     public static string projectSettingsPath => $"{SharedModule.SettingsEditor.projectSettingDirectory}/BuildAssist.json";
   }
 }