+++ Editor/Platform/BuildManagerPlatformAndroid.cs
@@ -1,13 +1,12 @@
 ﻿
-using Hananoki.Extensions;
+using HananokiRuntime.Extensions;
 using UnityEditor;
 using UnityEditor.Build.Reporting;
 using UnityEngine;
+using static HananokiEditor.BuildAssist.Console;
+using P = HananokiEditor.BuildAssist.SettingsProject;
 
-using static Hananoki.BuildAssist.Console;
-using P = Hananoki.BuildAssist.SettingsProject;
-
-namespace Hananoki.BuildAssist {
+namespace HananokiEditor.BuildAssist {
 
 	public class BuildPlatformAndroid : IBuildPlatform {