+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# BuildAssist
+# Build Assist
+
+## [1.3.7] - 2020-12-13
+- SharedModule v1.7.3 or later
+
+### Changed
+- Changed namespace
 
 ## [1.3.6] - 2020-12-02
 - SharedModule v1.7.0 or later